Move headers to src/ directory
