+++ examples/apriltags_demo.hpp
@@ -30,12 +30,12 @@
 #include "opencv2/opencv.hpp"
 
 // April tags detector and various families that can be selected by command line option
-#include "AprilTags/TagDetector.h"
-#include "AprilTags/Tag16h5.h"
-#include "AprilTags/Tag25h7.h"
-#include "AprilTags/Tag25h9.h"
-#include "AprilTags/Tag36h9.h"
-#include "AprilTags/Tag36h11.h"
+#include <apriltags/TagDetector.h>
+#include <apriltags/Tag16h5.h>
+#include <apriltags/Tag25h7.h>
+#include <apriltags/Tag25h9.h>
+#include <apriltags/Tag36h9.h>
+#include <apriltags/Tag36h11.h>
 
 // Needed for getopt / command line options processing
 #include <unistd.h>
