+++ src/FloatImage.h
