+++ src/GLineSegment2D.h
