+++ src/Gaussian.h
